fix(web): address CodeRabbit findings on settled banner
- Reset isUnsettling when the active thread changes so a mid-flight
  un-settle on thread A never renders thread B's banner as pending
  (mirrors the isRevertingCheckpoint reset pattern).
- Extract the duplicated minute-quantized clock into useNowMinute,
  shared by the sidebar partition and the composer banner.
